Fuse EinsplineSetBuilderOld.cpp into EinsplineSetBuilderESHDF.fft.cpp
